% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxds_bcds_hybrid.R
\name{cxds_bcds_hybrid}
\alias{cxds_bcds_hybrid}
\title{Find doublets/multiples in UMI scRNA-seq data;}
\usage{
cxds_bcds_hybrid(sce, cxds_args = NULL, bcds_args = NULL)
}
\arguments{
\item{sce}{single cell experiment to analyze; needs "counts" in assays slot.}

\item{cxds_args}{list; Arguments for cxds function in list form. Default: NULL}

\item{bcds_args}{list; Arguments for bcds function in list form. Default: NULL}
}
\value{
sce Input sce with doublet scores added to colData as "hybrid_score" column.
}
\description{
Find doublets/multiples in UMI scRNA-seq data;
}
