% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxds.R
\name{cxds_getTopPairs}
\alias{cxds_getTopPairs}
\title{Extract top-scoring gene pairs from an SingleCellExperiment where cxds has been run}
\usage{
cxds_getTopPairs(sce, n = 100)
}
\arguments{
\item{sce}{single cell experiment to analyze; needs "counts" in assays slot.}

\item{n}{integer. The number of gene pairs to extract. Default: 100}
}
\value{
matrix Matrix with two colulmns, each containing gene indexes for gene pairs (rows).
}
\description{
Extract top-scoring gene pairs from an SingleCellExperiment where cxds has been run
}
