% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_calls.R
\name{get_dblCalls_ALL}
\alias{get_dblCalls_ALL}
\title{Wrapper for getting doublet calls}
\usage{
get_dblCalls_ALL(scrs_real, scrs_sim, rel_loss = 1)
}
\arguments{
\item{scrs_real}{numeric vector, the scores for the real/original data}

\item{scrs_sim}{numeric vector, the scores for the artificial doublets}

\item{rel_loss}{numeric scalar, relative weight of a false positive classification compared with a false negative. Default:1 (same loss for fp and fn).}
}
\value{
numeric, matrix containing the (estimated) number of doublets, the score threshold and the fraction of artificial doublets missed (false negative rate, of sorts) as columns and four types of estimating: "youden", "balanced" and a false negative rate of artificial doublets of 0.1 and 0.01, respecitvely.
}
\description{
Wrapper for getting doublet calls
}
