% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce-data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Example single cell experiment (\code{SingleCellExperiment}) object}
\format{a single cell experiment object (\code{SingleCellExperiment}) with raw counts in the \code{counts} in assays, and colData with experimental annotations.}
\usage{
sce
}
\description{
Example data set, created by randomly sampling genes and cells from a real data set (ch_cl, i.e., the cell lines data from https://satijalab.org/seurat/hashing_vignette.html). Contains raw counts in the \code{counts} assay slot.
}
