% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcds.R
\name{bcds}
\alias{bcds}
\title{Find doublets/multiples in UMI scRNA-seq data;}
\usage{
bcds(sce, ntop = 500, srat = 1, verb = FALSE, retRes = FALSE,
  nmax = "tune", varImp = FALSE)
}
\arguments{
\item{sce}{single cell experiment to analyze; needs "counts" in assays slot.}

\item{ntop}{number of top variance genes to consider. Default: 500}

\item{srat}{ratio between orginal "cells" and simulated doubltes; Default:1}

\item{verb}{progress messages. Default: FALSE}

\item{retRes}{shuld the trained classifier be returned? Default: FALSE}

\item{nmax}{maximum number of training rounds; interger or "tune". Default: "tune"}

\item{varImp}{shold variable (i.e., gene) importance be returned? Default: FALSE}
}
\value{
sce Input sce with doublet scores added to colData as "bcds_score" column, and possibly more (details)
}
\description{
Find doublets/multiples in UMI scRNA-seq data;
}
