% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxds_bcds_hybrid.R
\name{cxds_bcds_hybrid}
\alias{cxds_bcds_hybrid}
\title{Find doublets/multiples in UMI scRNA-seq data;}
\usage{
cxds_bcds_hybrid(sce, cxds_args = NULL, bcds_args = NULL)
}
\arguments{
\item{sce}{single cell experiment (\code{SingleCellExperiment}) object to analyze; needs \code{counts} in assays slot.}

\item{cxds_args}{list, arguments for cxds function in list form. Default: NULL}

\item{bcds_args}{list, arguments for bcds function in list form. Default: NULL}
}
\value{
sce input sce object \code{SingleCellExperiment} with doublet scores added to colData as "hybrid_score" column.
}
\description{
Annotates doublets/multiplets using the hybrid approach
}
\examples{
data("sce_chcl")
## create small data set using only 100 cells
sce_chcl_small = sce_chcl[, 1:100]
sce_chcl_small = cxds_bcds_hybrid(sce_chcl_small)
}
